refactor(crypto): move crypto files into a separate folder
